clean up youtube resolver
